refactor: extractRoomId, extractName, getAuthentication 메서드 추가
