+++ backend/src/main/java/middle_point_search/backend/common/security/jwt/provider/JwtTokenProvider.java
@@ -1,14 +1,19 @@
-package middle_point_search.backend.common.security.jwt.service;
+package middle_point_search.backend.common.security.jwt.provider;
 
 import java.util.Optional;
+
+import org.springframework.security.core.Authentication;
 
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public interface JwtService {
+public interface JwtTokenProvider {
+
+	Authentication getAuthentication(String accessToken);
 
 	String createAccessToken(String roomId, String name);
+
 	String createRefreshToken();
 
 	void updateRefreshToken(String roomId, String name, String refreshToken);
@@ -16,6 +21,7 @@ public interface JwtService {
 	void destroyRefreshToken(String roomId, String name);
 
 	void sendAccessAndRefreshToken(HttpServletResponse response, String accessToken, String refreshToken);
+
 	void sendAccessToken(HttpServletResponse response, String accessToken);
 
 	Optional<String> extractAccessToken(HttpServletRequest request);
@@ -23,6 +29,10 @@ public interface JwtService {
 	Optional<String> extractRefreshToken(HttpServletRequest request);
 
 	Claims parseClaims(String accessToken);
+
+	Optional<String> extractName(String accessToken);
+
+	Optional<String> extractRoomId(String accessToken);
 
 	void setAccessTokenHeader(HttpServletResponse response, String accessToken);
